requirements.txt
pyyaml==6.0.2
anthropic==0.49.0
loguru==0.7.3
Pillow==12.1.1
PyAutoGUI==0.9.54
pyperclip==1.11.0
openai==1.75.0
paddleocr==3.4.0
paddlepaddle==3.3.1
